Add auth and signup validate middlewares
