+++ backend/middlewares/authenticate.js
@@ -1,0 +1,19 @@
+const jwt = require('jsonwebtoken');
+
+// Middleware for protected routes (example)
+const authenticateJWT = (req, res, next) => {
+    const token = req.header('Authorization')?.split(' ')[1];
+    if (token) {
+        jwt.verify(token, process.env.JWT_SECRET, (err, user) => {
+            if (err) {
+                return res.sendStatus(403); // Forbidden
+            }
+            req.user = user; // Attach user information to the request
+            next(); // Proceed to the next middleware or route handler
+        });
+    } else {
+        res.sendStatus(401); // Unauthorized
+    }
+};
+
+module.exports = authenticateJWT;
